requirements.txt
altair                                 
annotated-types                         
anyio                                   
asgiref                                 
asttokens                               
asyncio                                 
attrs                                   
Authlib                                 
backoff                                 
bcrypt                                  
blinker                                 
build                                   
cachetools                              
certifi                                  
cffi                                    
charset-normalizer                      
chroma-hnswlib                          
chromadb                                
chromadb-client                          
click                                   
colorama                                
coloredlogs                             
comm                                    
cryptography                            
cuda-python                             
debugpy                                 
decorator                               
Deprecated                              
dnspython                               
email_validator                         
exceptiongroup                          
executing                               
faiss-cpu                                
fastapi                                  
fastapi-cli                             
filelock                                
flatbuffers                              
fsspec                                   
gitdb                                   
GitPython                               
google-auth                             
googleapis-common-protos                
grpcio                                  
grpcio-health-checking                  
grpcio-tools                            
h11                                     
httpcore                                
httptools                               
httpx                                   
huggingface-hub                         
humanfriendly                           
idna                                    
importlib-metadata                      
importlib_resources                     
intel-openmp                             
ipykernel                               
ipython                                 
jedi                                    
Jinja2                                  
joblib                                  
jsonschema                              
jsonschema-specifications                
jupyter_client                          
jupyter_core                            
kubernetes                              
markdown-it-py                          
MarkupSafe                              
matplotlib-inline                       
mdurl                                   
mkl                                      
mmh3                                    
monotonic                               
mpmath                                  
nest_asyncio                            
networkx                                
nltk                                    
numpy                                   
oauthlib                                
onnxruntime                             
opentelemetry-api                       
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc  
opentelemetry-instrumentation           
opentelemetry-instrumentation-asgi      
opentelemetry-instrumentation-fastapi   
opentelemetry-proto                     
opentelemetry-sdk                       
opentelemetry-semantic-conventions      
opentelemetry-util-http                 
orjson                                  
overrides                               
packaging                               
pandas                                  
parso                                   
pickleshare                             
pillow                                  
pinecone-client                         
pip                                     
platformdirs                            
posthog                                 
prompt_toolkit                          
protobuf                                
psutil                                  
pure-eval                               
pyarrow                                 
pyasn1                                  
pyasn1_modules                          
pycparser                               
pydantic                                
pydantic_core                           
pydeck                                  
Pygments                                
PyPika                                  
pyproject_hooks                         
pyreadline3                             
python-dateutil                         
python-dotenv                           
python-multipart                        
pytz                                    
pywin32                                 
PyYAML                                  
pyzmq                                   
referencing                             
regex                                   
requests                                
requests-oauthlib                       
rich                                    
rpds-py                                 
rsa                                     
safetensors                             
scikit-learn                            
scipy                                   
sentence-transformers                   
setuptools                              
shellingham                             
six                                     
smmap                                   
sniffio                                 
stack-data                              
starlette                               
streamlit                               
sympy                                   
tbb                                     
tenacity                                
threadpoolctl                           
tokenizers                              
toml                                    
toolz                                   
torch                                   
tornado                                 
tqdm                                    
traitlets                               
transformers                            
typer                                   
typing_extensions                       
tzdata                                  
urllib3                                 
uvicorn                                 
validators                              
watchdog                                
watchfiles                              
wcwidth                                 
weaviate-client                         
websocket-client                        
websockets                              
wheel                                   
wrapt                                   
zipp                                    